Change color API to ints; rename scene; load lib
Switch Renderer2D.fill to use integer RGBA values and update JNI bindings accordingly: nativeFill now takes jint parameters and the JNI registration signature changed from (FFFF)V to (IIII)V. Update Kotlin external declaration to match (Int,Int,Int,Int) and default alpha to 255. Rename Scene1 to Scene3 and update MainActivity to register and set scene ID 3. Also ensure the native GraphicsEngine library is loaded by adding System.loadLibrary("GraphicsEngine") in MainActivity.init(). These changes align the Java/Kotlin API with the native signatures and update the active scene registration.

+++ app/src/main/java/edu/osu/t22/planear/GraphicsEngine.kt
@@ -1,7 +1,7 @@
 package edu.osu.t22.planear
 
 class Renderer2D(private val ptr: Long) {
-    external fun fill(r: Float, g: Float, b: Float, a: Float = 255.0f)
+    external fun fill(r: Int, g: Int, b: Int, a: Int = 255)
     external fun rect(x: Float, y: Float, width: Float, height: Float)
     external fun ellipse(x: Float, y: Float, width: Float, height: Float)
     external fun textFont(fontName: String, size: Int)
